chore: rename folder designSystem and files MultiSelect to match the naming convention
