+++ resources/css/app.css
@@ -1,4 +1,4 @@
-@import 'multiSelect.css';
+@import 'multi_select.css';
 
 .timeline {
   display: flex;
