+++ resources/css/multi_select.css
