+++ resources/js/app.js
@@ -1,5 +1,5 @@
 import '../css/app.css'
 import '../css/bs_override.css'
-import './multiSelect.js'
+import './multi_select.js'
 
 import.meta.glob(['../images/**', '../fonts/**'])
