+++ resources/views/design_system/design_system.edge
@@ -77,7 +77,7 @@
         </h2>
 
         <div class="col-12 col-md-6 col-lg-4">
-          @!component('components/multiSelect', {
+          @!component('components/multi_select', {
           name: 'Prévention',
           label: 'Prévention',
           options: ['Vol Simple', 'Menace de mort', 'Escroquerie', 'Fraude ficale', 'Délit de fuite'],
